Move locals middleware into routes
